Implement highlighting
continue previous work on highlighting:
make the response content available to the list
of results and result itself.
the default result component takes the hl results
for the current document, if some, and replace
the raw value (the one from the document) with it.
That means the value returned in hl must be the full
value + highlights tags, otherwise the value displayed
would be truncated.
The result components expect the highlight tags to be
some HTML since the hl result is parsed and transformed
to corresponding React component.
The dev project is updated accordingly with some config
into the hl node to make it work out of the box for books

Change-Id: I56ef41c87586bf2ae90b8235af000bfae3d583e3